+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `get-res-0-indexes` function for H3, which returns a collection of all indexes at resolution 0
 * Add `multi-point`, `multi-linestring`, and `multi-linestring-wkt` functions to `geo.jts`
 * Add testing support for JDK11 and Clojure 1.10
+* Fix reflection on geometry creation functions in the `jts` namespace
 * Bump `h3` to 3.4.0, enabling support for functions described above
 * Bump other core dependencies to keep up with upstream changes: `jts2geojson` to 0.13.0, and `jts` to 1.16.1
 * Bump internal dependencies for testing and documentation: `midje` to 1.9.6, `cheshire` to 5.8.1, and `lein-codox` to 0.10.6