+++ CHANGELOG.md
@@ -11,9 +11,9 @@
 * Add `multi-point`, `multi-linestring`, and `multi-linestring-wkt` functions to `geo.jts`
 * Add testing support for JDK11 and Clojure 1.10
 * Fix reflection on geometry creation functions in the `jts` namespace
-* Bump `h3` to 3.4.0, enabling support for functions described above
-* Bump other core dependencies to keep up with upstream changes: `jts2geojson` to 0.13.0, and `jts` to 1.16.1
-* Bump internal dependencies for testing and documentation: `midje` to 1.9.6, `cheshire` to 5.8.1, and `lein-codox` to 0.10.6
+* Bump `h3` to 3.4.1, enabling support for functions described above
+* Bump other core dependencies to keep up with upstream changes: `jts2geojson` to 0.14.0, and `jts` to 1.16.1
+* Bump internal dependencies for testing and documentation: `midje` to 1.9.9, `criterium` to 0.4.5, `cheshire` to 5.8.1, and `lein-codox` to 0.10.6
 * Add `deps.edn` based on `project.clj`
 
 ## 2.1.0 to 2.1.1