+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 * Add testing support for JDK11, JDK12, JDK13, and Clojure 1.10
 * Fix reflection on geometry creation functions in the `jts` namespace
 * Bump `h3` to 3.5.0, enabling support for functions described above
-* Bump other core dependencies to keep up with upstream changes: `jts2geojson` to 0.14.1, and `jts` to 1.16.1
+* Bump other core dependencies to keep up with upstream changes: `jts2geojson` to 0.14.2, and `jts` to 1.16.1
 * Bump internal dependencies for testing and documentation: `midje` to 1.9.9, `criterium` to 0.4.5, `cheshire` to 5.9.0, and `lein-codox` to 0.10.6
 * Add `deps.edn` based on `project.clj`